pyproject.toml
[project]
name = "orion-bot"
version = "0.1.0"
description = "Bot Telegram para controle remoto do PC com suporte a voz"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pycaw>=20240210",
    "comtypes>=1.4.0",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "yt-dlp>=2024.1.0",
    "pyautogui>=0.9.54",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
